[tls][tickets]: moving session_timeout to protobuf.Duration

+++ source/extensions/transport_sockets/tls/context_config_impl.h
@@ -134,7 +134,7 @@ public:
   const std::vector<SessionTicketKey>& sessionTicketKeys() const override {
     return session_ticket_keys_;
   }
-  uint32_t sessionTimeout() const override { return session_timeout_; }
+  uint32_t sessionTimeout() const override { return uint32_t(session_timeout_); }
 
   bool isReady() const override {
     const bool parent_is_ready = ContextConfigImpl::isReady();
@@ -161,7 +161,7 @@ private:
   getSessionTicketKeys(const envoy::api::v2::auth::TlsSessionTicketKeys& keys);
   ServerContextConfig::SessionTicketKey getSessionTicketKey(const std::string& key_data);
 
-  uint32_t session_timeout_{0};
+  uint64_t session_timeout_{0};
 };
 
 } // namespace Tls
